requirements.txt
paho-mqtt==1.5
requests==2.23